+++ strategy-game-backend/src/main/kotlin/de/ruegnerlukas/strategygame/backend/ports/models/dtos/TileDTO.kt
@@ -11,12 +11,16 @@ data class TileDTO(
 	val tileId: String,
 	val position: TilePosition,
 	val data: TileDTOData,
+	val influences: List<TileDTOCountryInfluence>,
+	val ownerCountryId: String?,
 	val content: List<TileDTOContent>
 ) {
 	constructor(tile: TileEntity) : this(
 		tile.key!!,
 		tile.position,
 		TileDTOData(tile.data.terrainType),
+		tile.influences.map { TileDTOCountryInfluence(it.countryId, it.value) },
+		tile.ownerCountryId,
 		tile.content.map {
 			when (it) {
 				is MarkerTileContent -> MarkerTileDTOContent(it.countryId)
@@ -28,6 +32,12 @@ data class TileDTO(
 
 data class TileDTOData(
 	var terrainType: String,
+)
+
+
+data class TileDTOCountryInfluence(
+	val countryId: String,
+	var value: Double,
 )