+++ strategy-game-frontend/src/core/actions/turn/turnUpdateWorldStateAction.ts
@@ -47,7 +47,12 @@ export class TurnUpdateWorldStateAction {
                     q: t.position.q,
                     r: t.position.r
                 },
-                terrainType: t.data.terrainType === "LAND" ? TerrainType.LAND : TerrainType.WATER
+                influences: t.influences.map(i => ({
+                    country: countries.find(c => c.countryId === i.countryId)!!,
+                    value: i.value
+                })),
+                terrainType: t.data.terrainType === "LAND" ? TerrainType.LAND : TerrainType.WATER,
+                ownerCountry: t.ownerCountryId ? countries.find(c => c.countryId === t.ownerCountryId) : undefined
             };
             t.content.forEach(c => {
                 if (c.type === "marker") {