+++ strategy-game-frontend/src/models/state/tile.ts
@@ -1,3 +1,4 @@
+import {Country} from "./country";
 import {TerrainType} from "./terrainType";
 import {TilePosition} from "./tilePosition";
 
@@ -5,4 +6,9 @@ export interface Tile {
     tileId: string,
     position: TilePosition;
     terrainType: TerrainType;
+    influences: ({
+        country: Country,
+        value: number,
+    })[],
+    ownerCountry: Country | undefined
 }