+++ strategy-game-frontend/src/shared/utils.ts
@@ -1,0 +1,7 @@
+export function orDefault<T>(value: T | null | undefined, defaultValue: T): T {
+    if (value == null || value == undefined) {
+        return defaultValue;
+    } else {
+        return value;
+    }
+}